tools: Restructure dtypes lowering
