+++ devito/tools/__init__.py
@@ -1,4 +1,5 @@
 from devito.tools.utils import *  # noqa
+from devito.tools.dtypes_lowering import *  # noqa
 from devito.tools.abc import *  # noqa
 from devito.tools.algorithms import *  # noqa
 from devito.tools.data_structures import *  # noqa
